+++ src/main/java/org/advent/year2019/intcode_computer/InputProvider.java
@@ -5,11 +5,11 @@ import java.util.Arrays;
 public interface InputProvider {
 	
 	boolean hasNext();
-	int nextInput();
+	long nextInput();
 	
 	
-	static InputProvider constant(int... values) {
-		int[] copy = Arrays.copyOf(values, values.length);
+	static InputProvider constant(long... values) {
+		long[] copy = Arrays.copyOf(values, values.length);
 		return new InputProvider() {
 			int index = 0;
 			
@@ -19,7 +19,7 @@ public interface InputProvider {
 			}
 			
 			@Override
-			public int nextInput() {
+			public long nextInput() {
 				if (hasNext())
 					return copy[index++];
 				throw new IllegalStateException("All input values are used");
@@ -38,7 +38,7 @@ public interface InputProvider {
 			}
 			
 			@Override
-			public int nextInput() {
+			public long nextInput() {
 				return copy[index++ % copy.length];
 			}
 		};
